Refactor text composition in controllers
