+++ easynav_tools/easynav_tools/tui/app.py
@@ -13,31 +13,12 @@ from textual.widgets import Static, Footer, Tabs, Tab, Label, Switch
 from textual.containers import Container, Horizontal, Vertical
 
 from ..controller.ros_controllers import (
-    TwistSubscriber,
-    TwistStampedSubscriber,
-    EasyNavControlSubscriber,
-    GoalManagerInfoSubscriber,
-    NavStateSubscriber
+    TwistProcessor,
+    TwistStampedProcessor,
+    EasyNavControlProcessor,
+    GoalManagerInfoProcessor,
+    NavStateProcessor
 )
-
-# Mapping of NavigationControl.type (uint8) to (label, color)
-NC_TYPE_MAP: dict[int, tuple[str, str]] = {
-    0: ("REQUEST",   "green"),
-    1: ("REJECT",    "red"),
-    2: ("ACCEPT",    "green"),
-    3: ("FEEDBACK",  "green"),
-    4: ("FINISHED",  "green"),
-    5: ("FAILED",    "red"),
-    6: ("CANCEL",    "yellow"),
-    7: ("CANCELLED", "yellow"),
-    8: ("ERROR",     "red"),
-}
-
-# Mapping for GoalManagerInfo.status (uint8)
-GM_STATUS_MAP: dict[int, tuple[str, str]] = {
-    0: ("IDLE",   "yellow"),
-    1: ("ACTIVE", "green"),
-}
 
 # -------- Running stats (Welford) --------
 class RunningStats:
@@ -170,12 +151,12 @@ class EasyNavTabbedApp(App):
 
         # ROS 2 subscribers
         self.subs = {}
-        self.subs["twist_subscriber"] = TwistSubscriber(self.node, self.twist_callback)
-        self.subs["twist_stamped_subscriber"] = TwistStampedSubscriber(
+        self.subs["twist_subscriber"] = TwistProcessor(self.node, self.twist_callback)
+        self.subs["twist_stamped_subscriber"] = TwistStampedProcessor(
             self.node, self.twist_stamped_callback
         )
-        self.subs["easynav_control"] = EasyNavControlSubscriber(self.node, self.control_callback)
-        self.subs["goal_info"] = GoalManagerInfoSubscriber(self.node, self.goal_info_callback)
+        self.subs["easynav_control"] = EasyNavControlProcessor(self.node, self.control_callback)
+        self.subs["goal_info"] = GoalManagerInfoProcessor(self.node, self.goal_info_callback)
 
         # Widget refs
         self.st_navstate: Static | None = None
@@ -283,7 +264,7 @@ class EasyNavTabbedApp(App):
         self.set_interval(0.05, self._ros_spin_once)
         # create NavState sub if switch is ON
         if self.query_one("#sw_navstate", Switch).value:
-            self.subs["navstate"] = NavStateSubscriber(self.node, self.navstate_callback)
+            self.subs["navstate"] = NavStateProcessor(self.node, self.navstate_callback)
         # Time stats: poll the log periodically (10 Hz is overkill; use ~2 Hz)
         self.set_interval(0.5, self._poll_time_stats_log)
 
@@ -318,7 +299,7 @@ class EasyNavTabbedApp(App):
             self.navstate_enabled = event.value
             if event.value:
                 # ON: (re)create subscriber and restore last content
-                self.subs["navstate"] = NavStateSubscriber(self.node, self.navstate_callback)
+                self.subs["navstate"] = NavStateProcessor(self.node, self.navstate_callback)
                 if self.st_navstate:
                     self.st_navstate.update(self._last_navstate_text)
             else:
@@ -348,28 +329,24 @@ class EasyNavTabbedApp(App):
         except (KeyboardInterrupt, ExternalShutdownException):
             pass
 
-    # ---------- Formatting helpers ----------
-    @staticmethod
-    def _fmt_duration(dur) -> str:
-        try:
-            sec = dur.sec
-            nsec = dur.nanosec
-            return f"{sec + nsec/1e9:.3f}s"
-        except Exception:
-            return str(dur)
+    def set_navstate_text(self, text: str) -> None:
+        self._last_navstate_text = text
+        if self.st_navstate is not None:
+            if self.navstate_enabled:
+                self.st_navstate.update(text)
+            else:
+                self.st_navstate.update("")
 
-    @staticmethod
-    def _fmt_pose(pose) -> str:
-        try:
-            p = pose.pose if hasattr(pose, "pose") else pose
-            px, py, pz = p.position.x, p.position.y, p.position.z
-            ox, oy, oz, ow = p.orientation.x, p.orientation.y, p.orientation.z, p.orientation.w
-            return (
-                f"pos=({px:.3f}, {py:.3f}, {pz:.3f}), "
-                f"quat=({ox:.3f}, {oy:.3f}, {oz:.3f}, {ow:.3f})"
-            )
-        except Exception:
-            return str(pose)
+    def set_time_stats_rows(self, rows) -> None:
+        table = self._render_time_stats_table(rows)
+        self._last_timestats_text = table
+        if self.st_timestats is not None:
+            if self.timestats_enabled:
+                self.st_timestats.update(table)
+            else:
+                self.st_timestats.update("")
+
+    # ---------- Formatting helpers ----------
 
     @staticmethod
     def _render_time_stats_table(rows) -> Text:
@@ -408,121 +385,32 @@ class EasyNavTabbedApp(App):
         cosy_cosp = 1.0 - 2.0 * (y * y + z * z)
         return math.atan2(siny_cosp, cosy_cosp)
 
-    def set_navstate_text(self, text: str) -> None:
-        self._last_navstate_text = text
-        if self.st_navstate is not None:
-            if self.navstate_enabled:
-                self.st_navstate.update(text)
-            else:
-                self.st_navstate.update("")
-
-    def set_time_stats_rows(self, rows) -> None:
-        table = self._render_time_stats_table(rows)
-        self._last_timestats_text = table
-        if self.st_timestats is not None:
-            if self.timestats_enabled:
-                self.st_timestats.update(table)
-            else:
-                self.st_timestats.update("")
-
     # ---------- ROS callbacks ----------
+
     def control_callback(self, msg) -> None:
         if self.box_nav_control is None:
             return
 
-        t_val: int = msg.type
-        label, color = NC_TYPE_MAP.get(t_val, (str(t_val), "white"))
-        type_line = f"[{color}]{label}[/{color}]"
-
-        pose_txt = self._fmt_pose(msg.current_pose) if msg.current_pose else "—"
-        nav_time_txt = self._fmt_duration(msg.navigation_time) if msg.navigation_time else "—"
-        eta_txt = self._fmt_duration(msg.estimated_time_remaining) if msg.estimated_time_remaining else "—"
-        dist_cov_txt = f"{msg.distance_covered:.3f} m" if isinstance(msg.distance_covered, (int, float)) else "—"
-        dist_goal_txt = f"{msg.distance_to_goal:.3f} m" if isinstance(msg.distance_to_goal, (int, float)) else "—"
-
-        text = (
-            f"Type: {type_line}\n"
-            f"Message: {msg.status_message}\n"
-            f"Current pose: {pose_txt}\n"
-            f"Navigation time: {nav_time_txt}\n"
-            f"ETA: {eta_txt}\n"
-            f"Distance covered: {dist_cov_txt}\n"
-            f"Distance to goal: {dist_goal_txt}"
-        )
+        text = EasyNavControlProcessor.msg2text(msg)
         self.box_nav_control.update(text)
 
     def goal_info_callback(self, msg) -> None:
         if self.box_goal_info is None:
             return
 
-        s_val: int = msg.status
-        s_label, s_color = GM_STATUS_MAP.get(s_val, (str(s_val), "white"))
-        status_line = f"Status: [{s_color}]{s_label}[/{s_color}]"
-
-        pos_ok = (msg.position_distance <= msg.position_tolerance)
-        ang_ok = (msg.angle_distance <= msg.angle_tolerance)
-
-        if pos_ok:
-            pos_line = (
-                f"[green]Position: distance={msg.position_distance:.3f} m "
-                f"/ tol={msg.position_tolerance:.3f} m[/green]"
-            )
-        else:
-            pos_line = (
-                f"Position: distance={msg.position_distance:.3f} m "
-                f"/ tol={msg.position_tolerance:.3f} m"
-            )
-
-        if pos_ok and ang_ok:
-            ang_line = (
-                f"[green]Angle: distance={msg.angle_distance:.3f} rad "
-                f"/ tol={msg.angle_tolerance:.3f} rad[/green]"
-            )
-        else:
-            ang_line = (
-                f"Angle: distance={msg.angle_distance:.3f} rad "
-                f"/ tol={msg.angle_tolerance:.3f} rad"
-            )
-
-        goals_count = len(msg.goals.goals)
-        goals_line = f"Goals remaining: {goals_count}"
-
-        if goals_count > 0:
-            g0 = msg.goals.goals[0]
-            pose0 = g0.pose if hasattr(g0, "pose") else g0
-            x = pose0.position.x
-            y = pose0.position.y
-            z = pose0.position.z
-            yaw = self._quat_to_yaw(pose0.orientation)
-            first_goal_line = f"First goal: x={x:.3f}, y={y:.3f}, z={z:.3f}, yaw={yaw:.3f} rad"
-        else:
-            first_goal_line = "First goal: —"
-
-        self.box_goal_info.update("\n".join([status_line, pos_line, ang_line, goals_line, first_goal_line]))
+        text = GoalManagerInfoProcessor.msg2text(msg)
+        self.box_goal_info.update(text)
 
     def twist_callback(self, msg) -> None:
-        self._last_twist_text = (
-            "Twist:\n"
-            f"  linear : x={msg.linear.x:.3f}, y={msg.linear.y:.3f}, z={msg.linear.z:.3f}\n"
-            f"  angular: x={msg.angular.x:.3f}, y={msg.angular.y:.3f}, z={msg.angular.z:.3f}"
-        )
+        self._last_twist_text = TwistProcessor.msg2text(msg)
         self._update_twist_box()
 
     def twist_stamped_callback(self, msg) -> None:
-        tw = msg.twist
-        self._last_twiststamped_text = (
-            "TwistStamped:\n"
-            f"  linear : x={tw.linear.x:.3f}, y={tw.linear.y:.3f}, z={tw.linear.z:.3f}\n"
-            f"  angular: x={tw.angular.x:.3f}, y={tw.angular.y:.3f}, z={tw.angular.z:.3f}"
-        )
-        if self.page_commanding:
-            self.page_commanding.update(
-                f"{self._last_twiststamped_text}\n\nSalir: 'q' o 'Ctrl+C'."
-            )
+        self._last_twiststamped_text = TwistStampedProcessor.msg2text(msg)
         self._update_twist_box()
 
     def navstate_callback(self, msg) -> None:
-        text = msg.data if hasattr(msg, "data") else str(msg)
+        text = NavStateProcessor.msg2text(msg)
         self._last_navstate_text = text
         if self.navstate_enabled and self.st_navstate is not None:
             self.st_navstate.update(text)
